requirements.txt
Flask-RESTful==0.3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
Pillow==10.1.0
PyMySQL==0.9.2
PyYAML==6.0.1
SQLAlchemy==1.2.12
XlsxWriter==3.2.0
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
coloredlogs==10.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
et-xmlfile==1.1.0
flask==3.0.2
fonttools==4.45.1
humanfriendly==10.0
importlib-metadata==7.1.0
importlib-resources==6.1.1
itsdangerous==2.1.2
jdcal==1.4.1
kiwisolver==1.4.5
matplotlib==3.8.2
more-itertools==10.2.0
mysql-connector-python==8.3.0
networkx==3.2.1
numpy==1.26.2
openpyxl==3.0.10
packaging==23.2
pandas==2.1.3
pip==22.3.1
plotly==5.18.0
pluggy==1.4.0
py==1.11.0
pycparser==2.21
pyparsing==3.1.1
pytest==3.8.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.4
setuptools==65.5.1
simplejson==3.19.2
six==1.16.0